refactor: break up game state entities moved game state components/entities into seperate files organized by concern
possible next step would be to seperate schema definitions from game
logic but this will be rather involved and requires a bit more thought
